requirements.txt
# Core dependencies
Flask>=3.0.0
Flask-CORS>=4.0.0
playwright>=1.40.0
asyncio>=3.4.3

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9

# Security
PyJWT==2.8.0
bcrypt==4.1.2
python-dotenv==1.0.0

# HTTP and API
requests==2.31.0
httpx==0.25.2
aiohttp==3.9.1

# Image and media processing
Pillow==10.1.0
ffmpeg-python==0.2.0

# Utilities
python-dateutil==2.8.2
pydantic==2.5.0
PyYAML==6.0.1

# Monitoring and logging
structlog==23.2.0
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# Development tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Performance
locust==2.17.0
memory-profiler==0.61.0

# Cache
redis==5.0.1